DihhJ Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
email-validator==2.1.0
